100k Message test ran
