--- K_Way_Merge_Kafka/KWayMerge/pom.xml
@@ -1,56 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <groupId>com.example.kafka</groupId>
-    <artifactId>KWayMerge</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    
-    <properties>
-        <kafka.version>3.7.0</kafka.version>
-	</properties>
-    
-    <dependencies>
-        <!-- Apache Kafka dependency -->
-        <dependency>
-            <groupId>org.apache.kafka</groupId>
-            <artifactId>kafka-clients</artifactId>
-            <version>${kafka.version}</version>
-        </dependency>
-      <dependency>
-        <groupId>junit</groupId>
-        <artifactId>junit</artifactId>
-        <version>4.13.2</version>
-        <scope>test</scope>
-      </dependency>
-
-	<dependency>
-		<groupId>org.testcontainers</groupId>
-		<artifactId>kafka</artifactId>
-		<version>1.19.7</version>
-		<scope>compile</scope>
-	</dependency>
-
-	<dependency>
-		<groupId>org.jfree</groupId>
-		<artifactId>jfreechart</artifactId>
-		<version>1.5.3</version> <!-- Check for the latest version -->
-	</dependency>
-
-    </dependencies>
-    
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
