+++ path-tracker-pipelines/src/main/java/org/example/KafkaConsumerThread.java
@@ -1,4 +1,22 @@
-package com.example.kafka;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.example;
 
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
@@ -39,7 +57,7 @@ public class KafkaConsumerThread implements Runnable{
         @Override
         public void run() {
             while (running) {
-                int pollTimeMilli = 1000;
+                int pollTimeMilli = 100;
                 ConsumerRecords<String, String> records = consumer.poll(Duration.ofMillis(
                         pollTimeMilli));
                 for (ConsumerRecord<String, String> record : records) {
@@ -51,7 +69,7 @@ public class KafkaConsumerThread implements Runnable{
                         partitionQueue[record.partition()].offer(message);
                     }
                     else {
-                        System.out.println("data is not in correct format closing");
+                        System.out.println("data is not in correct format closing " + record.value());
                         stopRunning();
                     }
                 }
